LayoutTests/http/tests/w3c/html/semantics/scripting-1/the-script-element/script-text-xhtml-expected.txt


FAIL Getter with CDATA section assert_equals: expected "\n\n\n  var x = \"y\";\n\n" but got "\n7;\n\n  var x = \"y\";\n\n"
